put most all parser setup in single module
