+++ conda/cli/main_install.py
@@ -6,73 +6,6 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from .conda_argparse import add_parser_create_install_update, add_parser_json
-from ..common.compat import on_win
-
-help = "Installs a list of packages into a specified conda environment."
-descr = help + """
-
-This command accepts a list of package specifications (e.g, bitarray=0.8)
-and installs a set of packages consistent with those specifications and
-compatible with the underlying environment. If full compatibility cannot
-be assured, an error is reported and the environment is not changed.
-
-Conda attempts to install the newest versions of the requested packages. To
-accomplish this, it may update some packages that are already installed, or
-install additional packages. To prevent existing packages from updating,
-use the --no-update-deps option. This may force conda to install older
-versions of the requested packages, and it does not prevent additional
-dependency packages from being installed.
-
-If you wish to skip dependency checking altogether, use the '--force'
-option. This may result in an environment with incompatible packages, so
-this option must be used with great caution.
-
-conda can also be called with a list of explicit conda package filenames
-(e.g. ./lxml-3.2.0-py27_0.tar.bz2). Using conda in this mode implies the
---force option, and should likewise be used with great caution. Explicit
-filenames and package specifications cannot be mixed in a single command.
-"""
-example = """
-Examples:
-
-    conda install -n myenv scipy
-
-"""
-
-
-def configure_parser(sub_parsers):
-    p = sub_parsers.add_parser(
-        'install',
-        description=descr,
-        help=help,
-        epilog=example,
-    )
-    p.add_argument(
-        "--revision",
-        action="store",
-        help="Revert to the specified REVISION.",
-        metavar='REVISION',
-    )
-    if on_win:
-        p.add_argument(
-            "--shortcuts",
-            action="store_true",
-            help="Install start menu shortcuts",
-            dest="shortcuts",
-            default=True
-        )
-        p.add_argument(
-            "--no-shortcuts",
-            action="store_false",
-            help="Don't install start menu shortcuts",
-            dest="shortcuts",
-            default=True
-        )
-    add_parser_create_install_update(p)
-    add_parser_json(p)
-    p.set_defaults(func=execute)
-
 
 def execute(args, parser):
     from .install import install
