+++ conda/cli/main_package.py
@@ -17,56 +17,8 @@ import re
 import tarfile
 import tempfile
 
-from .conda_argparse import add_parser_prefix
 from ..base.context import context
 from ..common.compat import PY3, itervalues
-
-descr = "Low-level conda package utility. (EXPERIMENTAL)"
-
-
-def configure_parser(sub_parsers):
-    p = sub_parsers.add_parser(
-        'package',
-        description=descr,
-        help=descr,
-    )
-    add_parser_prefix(p)
-    p.add_argument(
-        '-w', "--which",
-        metavar="PATH",
-        nargs='+',
-        action="store",
-        help="Given some PATH print which conda package the file came from.",
-    )
-    p.add_argument(
-        '-r', "--reset",
-        action="store_true",
-        help="Remove all untracked files and exit.",
-    )
-    p.add_argument(
-        '-u', "--untracked",
-        action="store_true",
-        help="Display all untracked files and exit.",
-    )
-    p.add_argument(
-        "--pkg-name",
-        action="store",
-        default="unknown",
-        help="Package name of the created package.",
-    )
-    p.add_argument(
-        "--pkg-version",
-        action="store",
-        default="0.0",
-        help="Package version of the created package.",
-    )
-    p.add_argument(
-        "--pkg-build",
-        action="store",
-        default=0,
-        help="Package build number of the created package.",
-    )
-    p.set_defaults(func=execute)
 
 
 def remove(prefix, files):
