+++ conda/cli/main_remove.py
@@ -6,108 +6,12 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from argparse import RawDescriptionHelpFormatter
 from collections import defaultdict
 import logging
 from os.path import isdir
 import sys
 
-from .conda_argparse import (add_parser_channels, add_parser_help, add_parser_insecure,
-                             add_parser_json, add_parser_no_pin, add_parser_offline,
-                             add_parser_prefix, add_parser_pscheck, add_parser_quiet,
-                             add_parser_use_index_cache, add_parser_use_local, add_parser_yes)
-
-help = "%s a list of packages from a specified conda environment."
-descr = help + """
-
-This command will also remove any package that depends on any of the
-specified packages as well---unless a replacement can be found without
-that dependency. If you wish to skip this dependency checking and remove
-just the requested packages, add the '--force' option. Note however that
-this may result in a broken environment, so use this with caution.
-"""
-example = """
-Examples:
-
-    conda %s -n myenv scipy
-
-"""
-
-uninstall_help = "Alias for conda remove.  See conda remove --help."
 log = logging.getLogger(__name__)
-
-
-def configure_parser(sub_parsers, name='remove'):
-    if name == 'remove':
-        p = sub_parsers.add_parser(
-            name,
-            formatter_class=RawDescriptionHelpFormatter,
-            description=descr % name.capitalize(),
-            help=help % name.capitalize(),
-            epilog=example % name,
-            add_help=False,
-        )
-    else:
-        p = sub_parsers.add_parser(
-            name,
-            formatter_class=RawDescriptionHelpFormatter,
-            description=uninstall_help,
-            help=uninstall_help,
-            epilog=example % name,
-            add_help=False,
-        )
-    add_parser_help(p)
-    add_parser_yes(p)
-    add_parser_json(p)
-    p.add_argument(
-        "--all",
-        action="store_true",
-        help="%s all packages, i.e., the entire environment." % name.capitalize(),
-    )
-    p.add_argument(
-        "--features",
-        action="store_true",
-        help="%s features (instead of packages)." % name.capitalize(),
-    )
-
-    # TODO: --features currently sorta still work. But super sloppy.
-
-    # p.add_argument(
-    #     '--feature',
-    #     metavar='FEATURE_NAME=FEATURE_VALUE',
-    #     dest='features',
-    #     action="append",
-    #     help="Feature to remove in the conda environment. "
-    #          "The value must be a key-value pair separated by an equal sign e.g. blas=nomkl. "
-    #          "Can be used multiple times. "
-    #          "Equivalent to a MatchSpec specifying a single 'provides_features'.",
-    # )
-
-    p.add_argument(
-        "--force",
-        action="store_true",
-        help="Forces removal of a package without removing packages that depend on it. "
-             "Using this option will usually leave your environment in a broken and "
-             "inconsistent state.",
-    )
-    add_parser_no_pin(p)
-    add_parser_channels(p)
-    add_parser_prefix(p)
-    add_parser_quiet(p)
-    # Putting this one first makes it the default
-    add_parser_use_index_cache(p)
-    add_parser_use_local(p)
-    add_parser_offline(p)
-    add_parser_pscheck(p)
-    add_parser_insecure(p)
-    p.add_argument(
-        'package_names',
-        metavar='package_name',
-        action="store",
-        nargs='*',
-        help="Package names to %s from the environment." % name,
-    )
-    p.set_defaults(func=execute)
 
 
 def execute(args, parser):
