+++ conda/common/configuration.py
@@ -29,7 +29,7 @@ from enum import Enum, EnumMeta
 
 from .compat import (isiterable, iteritems, itervalues, odict, primitive_types, string_types,
                      text_type, with_metaclass)
-from .constants import EMPTY_MAP, NULL
+from .constants import NULL
 from .path import expand
 from .serialize import yaml_load
 from .. import CondaError, CondaMultiError
@@ -54,6 +54,8 @@ except ImportError:  # pragma: no cover
     from ruamel.yaml.scanner import ScannerError
 
 log = getLogger(__name__)
+
+EMPTY_MAP = frozendict()
 
 
 def pretty_list(iterable, padding='  '):  # TODO: move elsewhere in conda.common
