+++ setup.cfg
@@ -24,13 +24,13 @@ markers =
 [pep8]
 max-line-length = 99
 ignore = E126,E133,E226,E241,E242,E302,E704,E731,W503
-exclude = build/*,.tox/*,tests/*,ve/*,*/_vendor/*,conda/compat.py,conda/common/compat.py,conda_env/compat.py
+exclude = build/*,env/*,.tox/*,tests/*,ve/*,*/_vendor/*,conda/compat.py,conda/common/compat.py,conda_env/compat.py
 
 
 [flake8]
 max-line-length = 99
 ignore = E126,E133,E226,E241,E242,E302,E704,E731,W503
-exclude = build/*,.tox/*,tests/*,ve/*,*/_vendor/*,conda/compat.py,conda/common/compat.py,conda_env/compat.py
+exclude = build/*,env/*,.tox/*,tests/*,ve/*,*/_vendor/*,conda/compat.py,conda/common/compat.py,conda_env/compat.py
 
 
 [coverage:report]
